+++ app/src/main/java/co/develhope/meteoapp/home/adapter/SubtitleViewHolder.kt
@@ -1,0 +1,11 @@
+package co.develhope.meteoapp.home.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import co.develhope.meteoapp.databinding.HomeSubtitleBinding
+import co.develhope.meteoapp.home.WeekItems.HomeSubtitle
+
+class SubtitleViewHolder(private val binding: HomeSubtitleBinding) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(item: HomeSubtitle) {
+
+    }
+}
