+++ app/src/main/java/co/develhope/meteoapp/home/adapter/TitleViewHolder.kt
@@ -1,0 +1,11 @@
+package co.develhope.meteoapp.home.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import co.develhope.meteoapp.databinding.HomeTitleBinding
+import co.develhope.meteoapp.home.WeekItems.HomeTitle
+
+class TitleViewHolder(private val binding: HomeTitleBinding) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(item: HomeTitle) {
+        // Qui manca il codice per settare il testo sulla text view, il luogo del meteo può cambiare quindi lo devi far arrivare fin qui
+    }
+}
