+++ app/src/main/java/co/develhope/meteoapp/home/setImageResource.kt
@@ -1,0 +1,7 @@
+package co.develhope.meteoapp.home
+
+import android.widget.ImageView
+import co.develhope.meteoapp.data.domain.DailySummaryForecast
+
+fun ImageView.setWeatherIcon(weatherIcon: DailySummaryForecast.WeatherIcon) {
+}
