+++ app/src/main/res/layout/fragment_home_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".HomeScreenFragment"
+    tools:context=".home.HomeScreenFragment"
     android:background="@drawable/gradient_background"
     >
