+++ app/src/main/res/layout/home_title.xml
@@ -2,20 +2,19 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
     <TextView
         android:id="@+id/home_title"
-        android:layout_width="235dp"
-        android:layout_height="39dp"
-        tools:text="@string/palermo_sicilia"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:text="@string/palermo_sicilia"
         android:textSize="32sp"
         android:layout_marginTop="42dp"
         android:layout_marginStart="24dp"
         android:textColor="@color/black"
         android:textFontWeight="600"
         android:fontFamily="@font/inter"
-        android:lineHeight="38.73dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
