+++ app/src/main/res/navigation/nav_graph.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/home_screen"
-        android:name="co.develhope.meteoapp.HomeScreenFragment"
+        android:name="co.develhope.meteoapp.home.HomeScreenFragment"
         android:label="fragment_home_screen"
         tools:layout="@layout/fragment_home_screen" />
     <fragment
